edit

+++ bot.py
@@ -24,12 +24,18 @@ async def on_ready():
         
 @client.event
 async def on_message(message):
+    if message.author == == client.user:
+        return
+    
     if pf.is_profane(message.content):
         await message.delete()
         msg = pf.censor(message.content)
         embed = discord.Embed(title="Profanity Detected!",description=msg, color=discord.Color.red())
         embed.set_author(name = message.author.display_name, icon_url = message.author.avatar_url)
         await message.channel.send(embed=embed)
+    
+    if message.channel.startswith("Hello"):
+        await message.channel.send("")
         
 @client.command()
 async def ping(ctx):
